Moved domain converter to separate file + added trianglemeshes and tetrahedralmeshes
